Moves `ImageCommentsUIComposer` to production
